separate script from transaction
